+++ pygeoapi/plugin.py
@@ -53,12 +53,12 @@ PLUGINS = {
         'MVT-tippecanoe': 'pygeoapi.provider.mvt_tippecanoe.MVTTippecanoeProvider',  # noqa: E501
         'MVT-elastic': 'pygeoapi.provider.mvt_elastic.MVTElasticProvider',
         'MVT-proxy': 'pygeoapi.provider.mvt_proxy.MVTProxyProvider',
-        'MySQL': 'pygeoapi.provider.mysql.MySQLProvider',
+        'MySQL': 'pygeoapi.provider.sql.MySQLProvider',
         'OracleDB': 'pygeoapi.provider.oracle.OracleProvider',
         'OGR': 'pygeoapi.provider.ogr.OGRProvider',
         'OpenSearch': 'pygeoapi.provider.opensearch_.OpenSearchProvider',
         'Parquet': 'pygeoapi.provider.parquet.ParquetProvider',
-        'PostgreSQL': 'pygeoapi.provider.postgresql.PostgreSQLProvider',
+        'PostgreSQL': 'pygeoapi.provider.sql.PostgreSQLProvider',
         'rasterio': 'pygeoapi.provider.rasterio_.RasterioProvider',
         'SensorThings': 'pygeoapi.provider.sensorthings.SensorThingsProvider',
         'SensorThingsEDR': 'pygeoapi.provider.sensorthings_edr.SensorThingsEDRProvider',  # noqa: E501
